requirements.txt
# Core dependencies for AgentMCP
# LLM and AI
openai>=1.0.0
anthropic>=0.7.0

# MCP Protocol
mcp>=1.0.0

# Web framework and UI
streamlit>=1.28.0
streamlit-chat>=0.1.1

# Data handling
pydantic>=2.0.0
jsonschema>=4.0.0
email-validator>=2.0.0

# HTTP and networking
requests>=2.31.0
httpx>=0.24.0

# Logging and monitoring
structlog>=23.0.0
python-json-logger>=2.0.0

# Configuration management
python-dotenv>=1.0.0
pydantic-settings>=2.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0
pytest-cov>=4.1.0

# Development tools
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
isort>=5.12.0

# Utilities
python-dateutil>=2.8.0
uuid>=1.30.0
typing-extensions>=4.7.0

# Optional: Database support (for future use)
# sqlalchemy>=2.0.0
# psycopg2-binary>=2.9.0
# pymongo>=4.4.0 